+++ src/PaginationExtension.ts
@@ -16,7 +16,8 @@ import { PaperOrientation, PaperSize } from "./types/paper";
 import { BorderConfig, MultiSide, MarginConfig } from "./types/page";
 import KeymapPlugin from "./Plugins/Keymap";
 import PaginationPlugin from "./Plugins/Pagination";
-import { getPageNodePosByPageNum, isPageNode } from "./utils/nodes/page";
+import { isPageNode } from "./utils/nodes/page/page";
+import { getPageNodePosByPageNum } from "./utils/nodes/page/pageNumber";
 import { isValidPaperSize, pageNodeHasPageSize, setPageNodePosPaperSize, setPagePaperSize } from "./utils/paperSize";
 import { getDeviceThemePaperColour, setPageNodePosPaperColour } from "./utils/paperColour";
 import { setBodyNodesAttribute, setPageNodesAttribute } from "./utils/setPageAttributes";
