+++ src/utils/getPageAttributes.ts
@@ -20,13 +20,14 @@ import { DEFAULT_PAPER_ORIENTATION } from "../constants/paperOrientation";
 import { DEFAULT_PAPER_COLOUR } from "../constants/paperColours";
 import { DEFAULT_PAGE_BORDER_CONFIG } from "../constants/pageBorders";
 import { PageRegionNodeAttributesObject } from "../types/pageRegions";
-import { doesDocHavePageNodes, getPageNodeByPageNum } from "./nodes/page";
+import { doesDocHavePageNodes } from "./nodes/page/page";
 import { FOOTER_DEFAULT_ATTRIBUTES, HEADER_DEFAULT_ATTRIBUTES } from "../constants/pageRegions";
 import { BODY_DEFAULT_ATTRIBUTES } from "../constants/body";
 import { getFooterNodeAttributes, getHeaderNodeAttributes } from "./pageRegion/pageRegion";
 import { getBodyNodeAttributes } from "./nodes/body";
 import { getPageRegionNode } from "./pageRegion/getAttributes";
 import { PaginationNodeAttributes } from "../types/pagination";
+import { getPageNodeByPageNum } from "./nodes/page/pageNumber";
 
 /**
  * Retrieves the page node attributes from the editor state.
