+++ src/utils/pageBorders.ts
@@ -12,7 +12,8 @@ import { DEFAULT_PAGE_BORDER_CONFIG } from "../constants/pageBorders";
 import { BorderConfig } from "../types/page";
 import { Nullable } from "../types/record";
 import { px } from "./units";
-import { getPageAttributeByPageNum } from "./nodes/page";
+import { getPageAttributeByPageNum } from "./nodes/page/pageNumber";
+
 import { setPageNodePosSideConfig, updatePageSideConfig } from "./setSideConfig";
 import { MultiAxisSide } from "../types/page";
