+++ src/utils/pageRegion/getAttributes.ts
@@ -11,9 +11,10 @@ import { PageRegion } from "../../types/pageRegions";
 import { Nullable } from "../../types/record";
 import { NullableNodePos } from "../../types/node";
 import { getStateFromContext } from "../editor";
-import { doesDocHavePageNodes, getPageNodeByPageNum, handleOutOfRangePageNum, isPageNumInRange } from "../nodes/page";
+import { doesDocHavePageNodes } from "../nodes/page/page";
 import { getHeaderFooterNodeType, isHeaderFooterNode } from "./pageRegion";
 import { isBodyNode } from "../nodes/body";
+import { getPageNodeByPageNum, handleOutOfRangePageNum, isPageNumInRange } from "../nodes/page/pageNumber";
 
 /**
  * Get the page region node of the current page by the page region type.
