+++ src/utils/paperSize.ts
@@ -13,7 +13,8 @@ import { DEFAULT_PAPER_ORIENTATION } from "../constants/paperOrientation";
 import { PaperOrientation, PaperDimensions, PaperSize } from "../types/paper";
 import { PageNodeAttributes, PageContentPixelDimensions } from "../types/page";
 import { Nullable } from "../types/record";
-import { getPageAttributeByPageNum, isPageNode } from "./nodes/page";
+import { isPageNode } from "./nodes/page/page";
+import { getPageAttributeByPageNum } from "./nodes/page/pageNumber";
 import { mmToPixels } from "./window";
 import { nodeHasAttribute } from "./attributes/getAttributes";
 import { setPageNodeAttribute } from "./setPageAttributes";
