+++ src/utils/setSideConfig.ts
@@ -10,9 +10,10 @@ import { Dispatch } from "@tiptap/core";
 import { pageSides } from "../constants/pageSides";
 import { PageSide, MultiSide, MultiAxisSide } from "../types/page";
 import { Nullable } from "../types/record";
-import { getPageNodePosByPageNum, isPageNode } from "./nodes/page";
+import { isPageNode } from "./nodes/page/page";
 import { setPageNodeAttribute } from "./setPageAttributes";
 import { NodePos } from "../types/node";
+import { getPageNodePosByPageNum } from "./nodes/page/pageNumber";
 
 type SideConfig<V> = { [key in PageSide]: V };
